Move tests from ConfigTest to ConfigFactoryJvmTest and then ConfigTest to shared
